added position subscriber to launch, fixed racetrack_evaluator

+++ racetrack_evaluator/racetrack_evelauator.py
@@ -3,12 +3,10 @@ import os.path
 
 import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
 from shapely.geometry import Point, Polygon
 from shapely.geometry.polygon import LinearRing, LineString
 import textwrap
 
-from deepracer.tracks import TrackIO, Track
 from deepracer.logs import \
     PlottingUtils as pu
 
@@ -95,7 +93,7 @@ def main():
   curve_3 = create_circle_angle((3.8, 1.0), points[3], np.pi/2-0.55, R_1, N_1)
   curve_4 = np.flipud(create_circle_angle((points[5,0],points[6,1]), points[6], -np.pi/2-0.55, R_1, N_1))
 
-  center_line = np.vstack([[0,2],curve_1,curve_2,curve_3,curve_4,[0,2]], dtype=np.float32)
+  center_line = np.vstack([[0,2],curve_1,curve_2,curve_3,curve_4,[0,2]])
   center_line = [line*6 + [-8.0, -8.0] for line in center_line]
   print(center_line)
   l_center_line = LineString(center_line)
